requirements.txt
python-telegram-bot==21.10
openai==1.12.0
httpx==0.27.2
httpcore==1.0.5
certifi==2024.12.14
h11==0.14.0
aiosqlite==0.20.0
aiohttp==3.9.5
tavily-python==0.7.21